调整字段组合模版upgrade时间
