+++ src/scene_server/admin_server/imports.go
@@ -88,6 +88,6 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202303301718"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202304012150"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202305110949"
-	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202305121010"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202305151505"
+	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202307141648"
 )
