+++ src/scene_server/admin_server/upgrader/y3.10.202307141648/add_field_template.go
@@ -15,7 +15,7 @@
  * to the current version of the project delivered to anyone in the future.
  */
 
-package y3_10_202305121010
+package y3_10_202307141648
 
 import (
 	"context"
